pyproject.toml
[project]
name = "race-performance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.19.0",
]
